// backend/relayer-server/typechain-types/MyForwarder.ts
export declare namespace MyForwarder {
  export type ForwardRequestStruct = {
    from: AddressLike;
    to: AddressLike;
    value: BigNumberish;
    gas: BigNumberish;
    deadline: BigNumberish;
    data: BytesLike;
    nonce: BigNumberish;
  };

  export type ForwardRequestStructOutput = [
    from: string,
    to: string,
    value: bigint,
    gas: bigint,
    deadline: bigint,
    data: string,
    nonce: bigint
  ] & {
    from: string;
    to: string;
    value: bigint;
    gas: bigint;
    deadline: bigint;
    data: string;
    nonce: bigint;
  };
}

export interface MyForwarderInterface extends Interface {
  getFunction(
    nameOrSignature: "execute" | "nonces" | "verify"
  ): FunctionFragment;

  encodeFunctionData(
    functionFragment: "execute",
    values: [MyForwarder.ForwardRequestStruct, BytesLike]
  ): string;
  encodeFunctionData(functionFragment: "nonces", values: [AddressLike]): string;
  encodeFunctionData(
    functionFragment: "verify",
    values: [MyForwarder.ForwardRequestStruct, BytesLike]
  ): string;

  decodeFunctionResult(functionFragment: "execute", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "nonces", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "verify", data: BytesLike): Result;
}

export interface MyForwarder extends BaseContract {
  connect(runner?: ContractRunner | null): MyForwarder;
  waitForDeployment(): Promise<this>;

  interface: MyForwarderInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  execute: TypedContractMethod<
    [req: MyForwarder.ForwardRequestStruct, signature: BytesLike],
    [string],
    "payable"
  >;

  nonces: TypedContractMethod<[arg0: AddressLike], [bigint], "view">;

  verify: TypedContractMethod<
    [req: MyForwarder.ForwardRequestStruct, signature: BytesLike],
    [boolean],
    "view"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "execute"
  ): TypedContractMethod<
    [req: MyForwarder.ForwardRequestStruct, signature: BytesLike],
    [string],
    "payable"
  >;
  getFunction(
    nameOrSignature: "nonces"
  ): TypedContractMethod<[arg0: AddressLike], [bigint], "view">;
  getFunction(
    nameOrSignature: "verify"
  ): TypedContractMethod<
    [req: MyForwarder.ForwardRequestStruct, signature: BytesLike],
    [boolean],
    "view"
  >;

  filters: {};
}